feateng/requirements.txt
zimply
python-baseconv
gensim
pandas
torch
torchvision
unidecode
spacy
openai
tqdm
nltk
bs4
scikit-learn
torchtext
faiss-cpu
